+++ src/PureHDF/API.Writing/H5SoftLink.cs
@@ -1,0 +1,7 @@
+namespace PureHDF;
+
+/// <summary>
+/// A soft link.
+/// </summary>
+/// <param name="Target">The link target path.</param>
+public record H5SoftLink(string Target);